+++ plugins/search-backend-module-techdocs/src/collators/DefaultTechDocsCollatorFactory.ts
@@ -22,6 +22,7 @@ import {
   CATALOG_FILTER_EXISTS,
   CatalogApi,
   CatalogClient,
+  EntityFilterQuery,
 } from '@backstage/catalog-client';
 import {
   Entity,
@@ -68,6 +69,8 @@ export type TechDocsCollatorFactoryOptions = {
   legacyPathCasing?: boolean;
   entityTransformer?: TechDocsCollatorEntityTransformer;
   documentTransformer?: TechDocsCollatorDocumentTransformer;
+  entityFilterFunction?: (entity: Entity[]) => Entity[];
+  customCatalogApiFilters?: EntityFilterQuery;
 };
 
 type EntityInfo = {
@@ -97,6 +100,8 @@ export class DefaultTechDocsCollatorFactory implements DocumentCollatorFactory {
   private readonly legacyPathCasing: boolean;
   private entityTransformer: TechDocsCollatorEntityTransformer;
   private documentTransformer: TechDocsCollatorDocumentTransformer;
+  private entityFilterFunction: Function | undefined;
+  private customCatalogApiFilters: EntityFilterQuery | undefined;
 
   private constructor(options: TechDocsCollatorFactoryOptions) {
     this.discovery = options.discovery;
@@ -110,6 +115,8 @@ export class DefaultTechDocsCollatorFactory implements DocumentCollatorFactory {
     this.legacyPathCasing = options.legacyPathCasing ?? false;
     this.entityTransformer = options.entityTransformer ?? (() => ({}));
     this.documentTransformer = options.documentTransformer ?? (() => ({}));
+    this.entityFilterFunction = options.entityFilterFunction;
+    this.customCatalogApiFilters = options.customCatalogApiFilters;
 
     this.auth = createLegacyAuthAdapters({
       auth: options.auth,
@@ -165,6 +172,7 @@ export class DefaultTechDocsCollatorFactory implements DocumentCollatorFactory {
             filter: {
               'metadata.annotations.backstage.io/techdocs-ref':
                 CATALOG_FILTER_EXISTS,
+              ...this.customCatalogApiFilters,
             },
             limit: batchSize,
             offset: entitiesRetrieved,
@@ -177,81 +185,71 @@ export class DefaultTechDocsCollatorFactory implements DocumentCollatorFactory {
       moreEntitiesToGet = entities.length === batchSize;
       entitiesRetrieved += entities.length;
 
-      const docPromises = entities
-        .filter(it => it.metadata?.annotations?.['backstage.io/techdocs-ref'])
-        .map((entity: Entity) =>
-          limit(async (): Promise<TechDocsDocument[]> => {
-            const entityInfo =
-              DefaultTechDocsCollatorFactory.handleEntityInfoCasing(
-                this.legacyPathCasing,
+      const filteredEntities = this.entityFilterFunction
+        ? this.entityFilterFunction(entities)
+        : this.defaultFilteringFunction(entities);
+
+      const docPromises = filteredEntities.map((entity: Entity) =>
+        limit(async (): Promise<TechDocsDocument[]> => {
+          const entityInfo =
+            DefaultTechDocsCollatorFactory.handleEntityInfoCasing(
+              this.legacyPathCasing,
+              {
+                kind: entity.kind,
+                namespace: entity.metadata.namespace || 'default',
+                name: entity.metadata.name,
+              },
+            );
+
+          try {
+            const { token: techdocsToken } =
+              await this.auth.getPluginRequestToken({
+                onBehalfOf: await this.auth.getOwnServiceCredentials(),
+                targetPluginId: 'techdocs',
+              });
+
+            const searchIndex = await fetch(
+              DefaultTechDocsCollatorFactory.constructDocsIndexUrl(
+                techDocsBaseUrl,
+                entityInfo,
+              ),
+              {
+                headers: {
+                  Authorization: `Bearer ${techdocsToken}`,
+                },
+              },
+            ).then(res => res.json());
+
+            return searchIndex.docs.map((doc: MkSearchIndexDoc) => ({
+              ...defaultTechDocsCollatorEntityTransformer(entity),
+              ...defaultTechDocsCollatorDocumentTransformer(doc),
+              ...this.entityTransformer(entity),
+              ...this.documentTransformer(doc),
+              location: this.applyArgsToFormat(
+                this.locationTemplate || '/docs/:namespace/:kind/:name/:path',
                 {
-                  kind: entity.kind,
-                  namespace: entity.metadata.namespace || 'default',
-                  name: entity.metadata.name,
+                  ...entityInfo,
+                  path: doc.location,
                 },
-              );
-
-            try {
-              const { token: techdocsToken } =
-                await this.auth.getPluginRequestToken({
-                  onBehalfOf: await this.auth.getOwnServiceCredentials(),
-                  targetPluginId: 'techdocs',
-                });
-
-              const searchIndexResponse = await fetch(
-                DefaultTechDocsCollatorFactory.constructDocsIndexUrl(
-                  techDocsBaseUrl,
-                  entityInfo,
-                ),
-                {
-                  headers: {
-                    Authorization: `Bearer ${techdocsToken}`,
-                  },
-                },
-              );
-
-              // todo(@backstage/techdocs-core): remove Promise.race() when node-fetch is 3.x+
-              // workaround for fetch().json() hanging in node-fetch@2.x.x, fixed in 3.x.x
-              // https://github.com/node-fetch/node-fetch/issues/665
-              const searchIndex = await Promise.race([
-                searchIndexResponse.json(),
-                new Promise((_resolve, reject) => {
-                  setTimeout(() => {
-                    reject('Could not parse JSON in 5 seconds.');
-                  }, 5000);
-                }),
-              ]);
-
-              return searchIndex.docs.map((doc: MkSearchIndexDoc) => ({
-                ...defaultTechDocsCollatorEntityTransformer(entity),
-                ...defaultTechDocsCollatorDocumentTransformer(doc),
-                ...this.entityTransformer(entity),
-                ...this.documentTransformer(doc),
-                location: this.applyArgsToFormat(
-                  this.locationTemplate || '/docs/:namespace/:kind/:name/:path',
-                  {
-                    ...entityInfo,
-                    path: doc.location,
-                  },
-                ),
-                ...entityInfo,
-                entityTitle: entity.metadata.title,
-                componentType: entity.spec?.type?.toString() || 'other',
-                lifecycle: (entity.spec?.lifecycle as string) || '',
-                owner: getSimpleEntityOwnerString(entity),
-                authorization: {
-                  resourceRef: stringifyEntityRef(entity),
-                },
-              }));
-            } catch (e) {
-              this.logger.debug(
-                `Failed to retrieve tech docs search index for entity ${entityInfo.namespace}/${entityInfo.kind}/${entityInfo.name}`,
-                e,
-              );
-              return [];
-            }
-          }),
-        );
+              ),
+              ...entityInfo,
+              entityTitle: entity.metadata.title,
+              componentType: entity.spec?.type?.toString() || 'other',
+              lifecycle: (entity.spec?.lifecycle as string) || '',
+              owner: getSimpleEntityOwnerString(entity),
+              authorization: {
+                resourceRef: stringifyEntityRef(entity),
+              },
+            }));
+          } catch (e) {
+            this.logger.debug(
+              `Failed to retrieve tech docs search index for entity ${entityInfo.namespace}/${entityInfo.kind}/${entityInfo.name}`,
+              e,
+            );
+            return [];
+          }
+        }),
+      );
       yield* (await Promise.all(docPromises)).flat();
     }
   }
@@ -265,6 +263,12 @@ export class DefaultTechDocsCollatorFactory implements DocumentCollatorFactory {
       formatted = formatted.replace(`:${key}`, value);
     }
     return formatted;
+  }
+
+  private defaultFilteringFunction(entities: Entity[]): Entity[] {
+    return entities.filter(
+      entity => entity.metadata?.annotations?.['backstage.io/techdocs-ref'],
+    );
   }
 
   private static constructDocsIndexUrl(