+++ packages/frontend-app-api/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @backstage/frontend-app-api
 
+## 0.11.0-next.2
+
+### Minor Changes
+
+- abcdf44: **BREAKING**: The returned object from `createSpecializedApp` no longer contains a `createRoot()` method, and it instead now contains `apis` and `tree`.
+
+  You can replace existing usage of `app.createRoot()` with the following:
+
+  ```ts
+  const root = tree.root.instance?.getData(coreExtensionData.reactElement);
+  ```
+
+- 8250ffe: **BREAKING**: Dropped support for the removed opaque `@backstage/ExtensionOverrides` and `@backstage/BackstagePlugin` types.
+
+### Patch Changes
+
+- 4d18b55: It's now possible to provide a middleware that wraps all extension factories by passing an `extensionFactoryMiddleware` to either `createApp()` or `createSpecializedApp()`.
+- Updated dependencies
+  - @backstage/frontend-defaults@0.2.0-next.2
+  - @backstage/frontend-plugin-api@0.10.0-next.2
+  - @backstage/core-app-api@1.16.0-next.0
+  - @backstage/config@1.3.2
+  - @backstage/core-plugin-api@1.10.4
+  - @backstage/errors@1.2.7
+  - @backstage/types@1.2.1
+  - @backstage/version-bridge@1.0.11
+
+## 0.10.6-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.2
+  - @backstage/core-app-api@1.15.5
+  - @backstage/core-plugin-api@1.10.4
+  - @backstage/errors@1.2.7
+  - @backstage/frontend-defaults@0.1.7-next.1
+  - @backstage/frontend-plugin-api@0.9.6-next.1
+  - @backstage/types@1.2.1
+  - @backstage/version-bridge@1.0.11
+
 ## 0.10.6-next.0
 
 ### Patch Changes