+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,43 @@
 # techdocs-cli-embedded-app
 
+## 0.2.106-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog@1.28.0-next.2
+  - @backstage/cli@0.31.0-next.1
+  - @backstage/core-app-api@1.16.0-next.0
+  - @backstage/core-components@0.16.5-next.1
+  - @backstage/app-defaults@1.6.0-next.1
+  - @backstage/test-utils@1.7.6-next.0
+  - @backstage/plugin-techdocs@1.12.4-next.2
+  - @backstage/plugin-techdocs-react@1.2.15-next.2
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/core-plugin-api@1.10.4
+  - @backstage/integration-react@1.2.5-next.0
+  - @backstage/theme@0.6.4
+
+## 0.2.106-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs@1.12.4-next.1
+  - @backstage/cli@0.30.1-next.0
+  - @backstage/core-components@0.16.5-next.0
+  - @backstage/plugin-catalog@1.28.0-next.1
+  - @backstage/app-defaults@1.5.18-next.0
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/core-app-api@1.15.5
+  - @backstage/core-plugin-api@1.10.4
+  - @backstage/integration-react@1.2.4
+  - @backstage/test-utils@1.7.5
+  - @backstage/theme@0.6.4
+  - @backstage/plugin-techdocs-react@1.2.15-next.1
+
 ## 0.2.106-next.0
 
 ### Patch Changes