+++ plugins/permission-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-permission-backend
 
+## 0.5.55-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-node@0.6.1-next.1
+  - @backstage/backend-plugin-api@1.2.1-next.1
+  - @backstage/config@1.3.2
+  - @backstage/errors@1.2.7
+  - @backstage/plugin-permission-common@0.8.4
+  - @backstage/plugin-permission-node@0.8.9-next.1
+
 ## 0.5.55-next.0
 
 ### Patch Changes