+++ plugins/proxy-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-proxy-node
 
+## 0.1.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.2.1-next.1
+
 ## 0.1.2-next.0
 
 ### Patch Changes