+++ plugins/scaffolder-backend-module-notifications/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-scaffolder-backend-module-notifications
 
+## 0.1.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.8.0-next.2
+  - @backstage/backend-plugin-api@1.2.1-next.1
+  - @backstage/plugin-notifications-common@0.0.8
+  - @backstage/plugin-notifications-node@0.2.13-next.2
+
+## 0.1.8-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.7.1-next.1
+  - @backstage/backend-plugin-api@1.2.1-next.1
+  - @backstage/plugin-notifications-common@0.0.8
+  - @backstage/plugin-notifications-node@0.2.13-next.1
+
 ## 0.1.8-next.0
 
 ### Patch Changes